chore: Prevent lockfile generation
We don't use lockfiles for our libraries, as `npm install` doesn't
consume anything other than the top-level lockfile anyway; this means
the tests continuously run on new versions, which means we catch issues
before they affect customers (hopefully).

The lockfile is current in the `.gitignore`, but may still be generated
locally, so this patch adds a `.npmrc` flag to prevent this.