chore: introduce Jest
This patch introduces Jest, and converts the remediation tests to use
it. It also removes a bunch of the virtualenv handling from the
remediation tests, as it seems superfluous. Finally, due to changing the
imports to stop casting as any, some additional typing was needed on the
test utils function defs