fix: ensure to always throw errors and suggest the cause
The plugin sometimes throws plain strings as the exception object, instead of an Error that wraps this string.
The resulting effect is that the Snyk CLI can work only with Error/exception objects and in such cases will print "undefined" which is a bad user experience.

The reason for a string error comes from the pip/python subprocess that we launch, so this fix also suggests to check if pip is installed in the current environment.